spinesim/gateway/requirements.txt
# VERTEX Gateway — Python dependencies
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
python-multipart==0.0.9
slowapi==0.1.9
prometheus-fastapi-instrumentator==7.0.0
stripe==10.0.0
